+++ src/components/ControlPlanner.vue
@@ -8,7 +8,9 @@
         <div>Selected faction:</div>
         <select v-model="plannerState.selectedFaction.value">
           <option :value="null">Abandoned</option>
-          <option :key="faction.key" v-for="faction in plannerState.plannerFactions" :value="faction">{{faction.name}}</option>
+          <optgroup :key="label" v-for="(factions, label) in plannerState.factionGroups" :label="label">
+            <option :key="faction.key" v-for="faction in factions" :value="faction">{{faction.name}}</option>
+          </optgroup>
         </select>
       </div>
 
@@ -17,6 +19,12 @@
         <button @click="plannerState.createBookmark">Copy link</button>
         <button @click="plannerState.reset">Reset</button>
         <button @click="plannerState.clear">Clear</button>
+
+        <div v-on="ioTooltipEvents">
+          WIP: more info <a href="https://steamcommunity.com/sharedfiles/filedetails/?id=2154921490" target="_blank" rel="noopener noreferrer">here</a>.
+        </div>
+        <button @click="plannerState.exportJson">Export (map.json)</button>
+        <button @click="plannerState.importJson">Import (map.json)</button>
       </div>
     </div>
   </div>
@@ -24,6 +32,7 @@
 
 <script>
 import { reactive, watch, toRefs, computed } from '@vue/composition-api';
+import { useTooltip } from '@/use/tooltip';
 
 export default {
   props: {
@@ -66,6 +75,9 @@ export default {
       }
     }, { lazy: true });
 
+    const { createTooltip } = useTooltip();
+    const ioTooltipEvents = createTooltip('text', 'Work in progress, please leave feedback on steam or github!');
+
     return {
       ...toRefs(state),
       collapse,
@@ -79,7 +91,8 @@ export default {
       stopPropagationListeners: {
         click: (e) => e.stopPropagation(),
         touchstart: (e) => e.stopPropagation()
-      }
+      },
+      ioTooltipEvents
     }
   }
 };